python/locust-test/requirements.txt
certifi==2019.6.16
chardet==3.0.4
Click==7.0
Flask==1.1.1
gevent==1.4.0
greenlet==0.4.15
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
locustio==0.11.0
MarkupSafe==1.1.1
msgpack==0.6.1
pyzmq==18.0.2
requests==2.22.0
six==1.12.0
urllib3==1.25.3
Werkzeug==0.15.5
